Elasticsearch/requirements.txt
elasticsearch==7.15.1
kafka-python==2.0.2
